Make a par_fold piece parse as its bytes do sequentially; fix two more
The frame check's promise is that split + parse + merge cannot disagree
with the sequential parse on a grammar it accepts. It could. The entry
point of every generated `parse_<RULE>()` skips whitespace before and
after the rule, and a `par_fold` rule's parser is the per-piece parser -
so that skip ran once per piece instead of once per input. On the
flagship grammar shape, `NAME = until(";")` under a "\n" frame:

    "A;\n B;\n  C;\n"   sequential 6   in two pieces 4

The piece starting at " B" lost its leading space to the wrapper; in one
go that space is part of the name. The mirror case is whitespace-only
text between two frames: the piece that ends there swallows it after the
fold and succeeds, the sequential parse has more input behind it and
fails. Trailing whitespace is not safe either, for that reason.

A `par_fold` rule's entry point now skips nothing on either side. The
sequential parse is stricter for it - a trailing blank line is an error
where it was tolerated - and that is the point: the two parses now agree
on every input, which is the property that makes cutting sound. The
whole-body rule already said a prefix or suffix belongs to no piece; the
wrapper's whitespace was such a prefix. A test runs eight awkward inputs
against 1..8 pieces and asserts agreement, on failure as well as on the
value.

`#[frame]` after a rule without an action block did not parse: `#` after
a rule body is also how a label (`#"text"`) begins, and the label parser
took the `#` and then demanded a string literal. Every example on the
branch happened to put an action on the preceding rule. A label is now
recognised by `#` followed by a string literal, so `#[…]` is left for
the next rule's attributes.

`frames()` computed the nominal start as `len * k / n` in `usize`, which
overflows before `len` does on a 32-bit target; the product is taken in
`u128`.

+++ src/rt.rs
@@ -130,7 +130,9 @@ pub fn frames(input: &str, boundary: &str, n: usize) -> Vec<std::ops::Range<usiz
     let mut starts = Vec::with_capacity(n + 1);
     starts.push(0);
     for k in 1..n {
-        let nominal = len * k / n;
+        // `len * k` in `usize` would overflow before `len` does on a 32-bit
+        // target; the quotient itself is at most `len`.
+        let nominal = ((len as u128 * k as u128) / n as u128) as usize;
         let tail: &[u8] = &bytes[nominal..];
         let start = tail
             .find_slice(boundary.as_bytes())
